Updated the logout and login
