+++ frontend/app/src/main/java/com/example/smartstorageorganizer/LoginActivity.java
@@ -58,7 +58,8 @@ public class LoginActivity extends AppCompatActivity {
                     finish();
 
                 } else {
-                    Log.i("AmplifyQuickstart", "User is not signed in");
+                    Toast.makeText(this, "User is not signed in.", Toast.LENGTH_LONG).show();
+                    Log.i("AmplifyQuickstart", "User is not signed in.");
                 }
             });
         signUpLink = findViewById(R.id.signUpLink);
@@ -179,34 +180,6 @@ public class LoginActivity extends AppCompatActivity {
 
 
     }
-
-
-    public void SignOut()
-    {
-        Amplify.Auth.signOut( signOutResult -> {
-            if (signOutResult instanceof AWSCognitoAuthSignOutResult.CompleteSignOut) {
-                // Sign Out completed fully and without errors.
-                Log.i("AuthQuickStart", "Signed out successfully");
-                //move to a different page
-            } else if (signOutResult instanceof AWSCognitoAuthSignOutResult.PartialSignOut) {
-                // Sign Out completed with some errors. User is signed out of the device.
-                AWSCognitoAuthSignOutResult.PartialSignOut partialSignOutResult =
-                        (AWSCognitoAuthSignOutResult.PartialSignOut) signOutResult;
-                //move to the different page
-
-            } else if (signOutResult instanceof AWSCognitoAuthSignOutResult.FailedSignOut) {
-                AWSCognitoAuthSignOutResult.FailedSignOut failedSignOutResult =
-                        (AWSCognitoAuthSignOutResult.FailedSignOut) signOutResult;
-                // Sign Out failed with an exception, leaving the user signed in.
-                Log.e("AuthQuickStart", "Sign out Failed", failedSignOutResult.getException());
-                //dont move to different page
-            }
-        });
-
-    }
-
-
-
 }
