+++ frontend/app/src/main/java/com/example/smartstorageorganizer/ProfileManagementActivity.java
@@ -2,9 +2,11 @@ package com.example.smartstorageorganizer;
 
 import android.content.Intent;
 import android.os.Bundle;
+import android.util.Log;
 import android.view.View;
 import android.widget.ImageView;
 import android.widget.TextView;
+import android.widget.Toast;
 
 import androidx.activity.EdgeToEdge;
 import androidx.appcompat.app.AppCompatActivity;
@@ -13,6 +15,9 @@ import androidx.constraintlayout.widget.ConstraintLayout;
 import androidx.core.graphics.Insets;
 import androidx.core.view.ViewCompat;
 import androidx.core.view.WindowInsetsCompat;
+
+import com.amplifyframework.auth.cognito.result.AWSCognitoAuthSignOutResult;
+import com.amplifyframework.core.Amplify;
 
 public class ProfileManagementActivity extends AppCompatActivity {
 
@@ -45,9 +50,7 @@ public class ProfileManagementActivity extends AppCompatActivity {
         logoutButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                Intent intent = new Intent(ProfileManagementActivity.this, LoginActivity.class);
-                startActivity(intent);
-                finish();
+                SignOut();
             }
         });
 
@@ -58,5 +61,43 @@ public class ProfileManagementActivity extends AppCompatActivity {
                 startActivity(intent);
             }
         });
+    }
+
+    public void SignOut()
+    {
+        Amplify.Auth.signOut(signOutResult -> {
+            if (signOutResult instanceof AWSCognitoAuthSignOutResult.CompleteSignOut) {
+                // Sign Out completed fully and without errors.
+                Log.i("AuthQuickStart", "Signed out successfully");
+                //move to a different page
+                runOnUiThread(() -> {
+                    Intent intent = new Intent(ProfileManagementActivity.this, LoginActivity.class);
+                    startActivity(intent);
+                    finish();
+                });
+            } else if (signOutResult instanceof AWSCognitoAuthSignOutResult.PartialSignOut) {
+                // Sign Out completed with some errors. User is signed out of the device.
+                AWSCognitoAuthSignOutResult.PartialSignOut partialSignOutResult =
+                        (AWSCognitoAuthSignOutResult.PartialSignOut) signOutResult;
+                //move to the different page
+                runOnUiThread(() -> {
+                    Intent intent = new Intent(ProfileManagementActivity.this, LoginActivity.class);
+                    startActivity(intent);
+                    finish();
+                });
+
+            } else if (signOutResult instanceof AWSCognitoAuthSignOutResult.FailedSignOut) {
+                AWSCognitoAuthSignOutResult.FailedSignOut failedSignOutResult =
+                        (AWSCognitoAuthSignOutResult.FailedSignOut) signOutResult;
+                // Sign Out failed with an exception, leaving the user signed in.
+                Log.e("AuthQuickStart", "Sign out Failed", failedSignOutResult.getException());
+                //dont move to different page
+                runOnUiThread(() -> {
+                    Toast.makeText(this, "Sign Out Failed.", Toast.LENGTH_LONG).show();
+
+                });
+            }
+        });
+
     }
 }
